requirements.txt
transformers
python-dotenv
tensorflow
pillow